+++ src/components/live_feed/comment.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import moment from 'moment';
-import './comment.scss';
 
 class Comment extends React.Component {
   constructor() {
