+++ src/components/live_feed/comment_form.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import './comment.scss';
 
 class CommentForm extends React.Component {
   constructor() {
