+++ src/style/style.scss
@@ -7,6 +7,7 @@
 @import "./upload_image.scss";
 @import "./event-page.scss";
 @import "./filter-container.scss";
+@import "./comment.scss";
 
 html, body {
   font-family: 'Roboto', 'Old Standard TT', 'Times New Roman', Times, serif;
